Remove additional duplicates relating to refactoring.
- Also remove BodyUpdater & CollectingSearchRequestor
