+++ org.eclipse.jdt.ls.tests/src/org/eclipse/jdt/ls/core/internal/handlers/ChangeSignatureHandlerTest.java
@@ -28,12 +28,12 @@ import org.eclipse.jdt.core.IPackageFragment;
 import org.eclipse.jdt.core.IPackageFragmentRoot;
 import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.core.JavaModelException;
+import org.eclipse.jdt.internal.corext.refactoring.ParameterInfo;
 import org.eclipse.jdt.internal.corext.util.JdtFlags;
 import org.eclipse.jdt.ls.core.internal.CodeActionUtil;
 import org.eclipse.jdt.ls.core.internal.JavaClientConnection;
 import org.eclipse.jdt.ls.core.internal.JavaCodeActionKind;
 import org.eclipse.jdt.ls.core.internal.LanguageServerWorkingCopyOwner;
-import org.eclipse.jdt.ls.core.internal.corext.refactoring.ParameterInfo;
 import org.eclipse.jdt.ls.core.internal.handlers.ChangeSignatureHandler.MethodException;
 import org.eclipse.jdt.ls.core.internal.handlers.ChangeSignatureHandler.MethodParameter;
 import org.eclipse.jdt.ls.core.internal.text.correction.RefactorProposalUtility;
@@ -94,7 +94,7 @@ public class ChangeSignatureHandlerTest extends AbstractCompilationUnitBasedTest
 		Assert.assertEquals(3, arguments.size());
 		Object arg0 = arguments.get(0);
 		assertEquals(true, arg0 instanceof String);
-		assertEquals("changeSignature", (String) arg0);
+		assertEquals("changeSignature", arg0);
 		Object arg1 = arguments.get(1);
 		assertEquals(true, arg1 instanceof CodeActionParams);
 		Object arg2 = arguments.get(2);
